lxd/backup: Moves Info struct and GetInfo function into own file
